+++ src/main/java/com/salcedoFawcett/services/clientService/persistence/entity/PartyIdentification.java
@@ -8,7 +8,7 @@ public class PartyIdentification {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-    private double partyIdentificationId;
+    private long partyIdentificationId;
     private int partyIdentificationSchemeId;
     private int partyIdentificationSchemeName;
 
@@ -23,11 +23,11 @@ public class PartyIdentification {
         this.id = id;
     }
 
-    public double getPartyIdentificationId() {
+    public long getPartyIdentificationId() {
         return partyIdentificationId;
     }
 
-    public void setPartyIdentificationId(double partyIdentificationId) {
+    public void setPartyIdentificationId(long partyIdentificationId) {
         this.partyIdentificationId = partyIdentificationId;
     }
